requirements.txt
boto3==1.42.9
botocore==1.42.9
jmespath==1.0.1
PyMuPDF==1.26.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.2